Cliente para API do Spotify em serviço à parte - Desmembramento de SpotifyPlaylistGetter
